+++ CHANGELOG.md
@@ -4,12 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Enabled logging and log aggregation ([#547]).
+
 ### Changed
 
 - operator-rs: 0.30.1 -> 0.33.0 ([#545]).
 - Bumped stackable versions to "23.4.0-rc1" ([#545]).
+- Bumped kafka stackable versions to "23.4.0-rc2" ([#547]).
 
 [#545]: https://github.com/stackabletech/kafka-operator/pull/545
+[#547]: https://github.com/stackabletech/kafka-operator/pull/547
 
 ## [23.1.0] - 2023-01-23